requirements.txt
gradio==5.23.3
gradio-webrtc[vad]==0.0.17
google-genai==1.3.0
python-dotenv==1.0.1
pillow==10.4.0
